Moved the loggers to a separate library.
