+++ LogLibrary/LogTarget.cs
