+++ MusicBrainzApp.sln
@@ -5,9 +5,11 @@ VisualStudioVersion = 17.0.32112.339
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MusicBrainzConsoleApp", "MusicBrainzConsoleApp\MusicBrainzConsoleApp.csproj", "{E7D48919-9753-4D2B-B59C-025B8B6FBCF4}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MusicBrainzEntitiesLibrary", "MusicBrainzEntitiesLibrary\MusicBrainzEntitiesLibrary.csproj", "{5D6F8C45-4E91-4251-9BA7-B0D81EE82488}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MusicBrainzEntitiesLibrary", "MusicBrainzEntitiesLibrary\MusicBrainzEntitiesLibrary.csproj", "{5D6F8C45-4E91-4251-9BA7-B0D81EE82488}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MusicBrainzTestLibrary", "MusicBrainzTestLibrary\MusicBrainzTestLibrary.csproj", "{AD9B0F40-F949-45E6-BD12-2E927DC11F20}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MusicBrainzTestLibrary", "MusicBrainzTestLibrary\MusicBrainzTestLibrary.csproj", "{AD9B0F40-F949-45E6-BD12-2E927DC11F20}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LogLibrary", "LogLibrary\LogLibrary.csproj", "{0B1BDCC8-14E6-4E14-9D8C-96712BA81816}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +29,10 @@ Global
 		{AD9B0F40-F949-45E6-BD12-2E927DC11F20}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AD9B0F40-F949-45E6-BD12-2E927DC11F20}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AD9B0F40-F949-45E6-BD12-2E927DC11F20}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0B1BDCC8-14E6-4E14-9D8C-96712BA81816}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0B1BDCC8-14E6-4E14-9D8C-96712BA81816}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0B1BDCC8-14E6-4E14-9D8C-96712BA81816}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0B1BDCC8-14E6-4E14-9D8C-96712BA81816}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
